ZOOKEEPER-2924: Refactor tests of LoadFromLogTest.java
This patch is the trunk version of https://github.com/apache/zookeeper/pull/409

Author: Andor Molnár <andor@cloudera.com>

Reviewers: phunt@apache.org

Closes #420 from anmolnar/ZOOKEEPER-2924-trunk
